상품 리뷰 작성 후 리프레시 로직 수정

+++ app/src/main/java/io/temco/guhada/data/model/InquiryRequest.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.Expose
 
 class InquiryRequest {
     @Expose
-    var productId: Int = 0
+    var productId: Long = 0
 
     var content: String = ""
     var privateInquiry: Boolean = false
